third_party/WebKit/LayoutTests/plugins/navigator-pluginarray-deleteframe.html

<!doctype html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script type="text/javascript">
test(function () {
  var iframe = document.createElement('iframe');
  document.body.appendChild(iframe);
  var iframe_navigator = iframe.contentWindow.navigator;
  assert_greater_than_equal(iframe_navigator.plugins.length, 1, "At least one plugin must be available.");
  assert_greater_than_equal(iframe_navigator.mimeTypes.length, 1, "At least one mime type must be available.");
  iframe.remove();
  for (var i = 0; i < iframe_navigator.plugins.length; i++) {
    assert_equals(null, iframe_navigator.plugins[i]);
  }
  for (var i = 0; i < iframe_navigator.mimeTypes.length; i++) {
    assert_equals(null, iframe_navigator.mimeTypes[i]);
  }
}, "Tests that navigator.plugins returns null once the iframe is removed.");
</script>
</body>
</html>
